backend/requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
python-multipart
python-docx
